+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
-- geyser: use runtime instaed of unconstrained ([#331](https://github.com/rpcpool/yellowstone-grpc/pull/331))
+- geyser: use runtime instaed of unconstrained ([#332](https://github.com/rpcpool/yellowstone-grpc/pull/332))
 
 ### Breaking